+++ benches/test_benches.rs
@@ -6,10 +6,11 @@ use std::time::{Duration, Instant};
 
 const TIMEOUT: u64 = 30;
 
-const HEURISTICS: [Heuristic; 5] = [
+const HEURISTICS: [Heuristic; 6] = [
     Heuristic::LeftBiased,
     Heuristic::Interleave,
     Heuristic::SmallFirst,
+    Heuristic::Hybrid,
     Heuristic::LookAhead,
     Heuristic::Random,
 ];